+++ crates/core/src/sources/silence.rs
@@ -4,8 +4,8 @@ use crate::{NextState, RawedioError, Sound};
 
 /// A forever stream of samples of value 0 (creating no sound).
 pub struct Silence {
-    channel_count: u16,
-    sample_rate: u32,
+    channel_count: usize,
+    sample_rate: usize,
 }
 
 impl Silence {
@@ -13,7 +13,7 @@ impl Silence {
     /// return the specified channel count and sample rate from
     /// their respective methods.
     #[must_use]
-    pub const fn new(channel_count: u16, sample_rate: u32) -> Silence {
+    pub const fn new(channel_count: usize, sample_rate: usize) -> Silence {
         Silence {
             channel_count,
             sample_rate,
@@ -22,16 +22,16 @@ impl Silence {
 }
 
 impl Sound for Silence {
-    fn channel_count(&self) -> u16 {
+    fn channel_count(&self) -> usize {
         self.channel_count
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> usize {
         self.sample_rate
     }
 
-    fn fill_next_frames(&mut self, buffer: &mut [i16]) -> Result<(usize, NextState), RawedioError> {
-        buffer.fill(0);
+    fn fill_next_frames(&mut self, buffer: &mut [f32]) -> Result<(usize, NextState), RawedioError> {
+        buffer.fill(0.0);
         Ok((buffer.len(), NextState::Playing))
     }
 }
